requirements.txt
pandas==1.3.3
nltk==3.4.5
tqdm==4.41.1
sqlalchemy==1.3.20
lxml==4.6.1
requests==2.22.0
urllib3==1.25.11
beautifulsoup4==4.9.3
pymystem3==0.2.0
pymorphy2==0.9.1
py==1.9.0
selenium==3.141.0
datedelta==1.3
tidytext==0.0.1
tensorflow==2.5.0
keras==2.5.0rc0
scikit-learn==1.0
psycopg2==2.9.3
numpy==1.19.5
cloudscraper==1.2.58
spacy==3.2.2
navec==0.10.0
PyYAML==6.0
feedly-client==0.23.2
openpyxl==3.0.9
xgboost==1.4.2